move css to where it will be included in dist
